requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.53.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asyncstdlib-fw==3.13.2
attrs==25.3.0
beautifulsoup4==4.13.4
betterproto-fw==2.0.3
blinker==1.9.0
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dataclasses-json==0.6.7
datasets==3.6.0
deep-sort-realtime==1.3.2
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
faiss-cpu==1.11.0
fastapi==0.115.13
ffmpeg-python==0.2.0
filelock==3.18.0
fireworks-ai==0.17.18
fonttools==4.58.4
frozenlist==1.6.2
fsspec==2025.3.0
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.32.4
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.65
langchain-fireworks==0.3.0
langchain-huggingface==0.2.0
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mcp==1.9.2
mdurl==0.1.2
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.42.1
networkx==3.5
numpy==2.3.0
obsws-python==1.7.1
omegaconf==2.3.0
openai==1.88.0
opencv-python==4.11.0.86
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
pandera==0.24.0
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
postgrest==1.1.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.7.0
PyPDF2==3.0.1
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
realtime==2.5.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-cache==1.2.1
requests-toolbelt==1.0.0
reward-kit==0.3.4
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.9.10
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==79.0.1
shellingham==1.5.4
shiboken6==6.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
statsbombpy==1.16.0
storage3==0.12.0
streamlit==1.45.1
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchreid==0.2.5
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typeguard==4.4.4
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.160
ultralytics-thop==2.0.14
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.0
zstandard==0.23.0